Fix Build

+++ orttraining/orttraining/core/optimizer/localized_recompute.cc
@@ -27,7 +27,7 @@ Status GeluRecompute::ApplyImpl(Graph& graph, bool& modified, int /*graph_level*
   GraphViewer graph_viewer(graph);
   const auto& node_ids = graph_viewer.GetNodesInTopologicalOrder();
 
-  for (int i = node_ids.size() - 1; i >= 0; --i) {
+  for (size_t i = node_ids.size() - 1; i >= 0; --i) {
     Node& node = *graph.GetNode(node_ids[i]);
 
     if (!SatisfyCondition(node)) {
@@ -72,7 +72,7 @@ Status AttentionDropoutRecompute::ApplyImpl(Graph& graph, bool& modified, int /*
   GraphViewer graph_viewer(graph);
   const auto& node_ids = graph_viewer.GetNodesInTopologicalOrder();
 
-  for (int i = node_ids.size() - 1; i >= 0; --i) {
+  for (size_t i = node_ids.size() - 1; i >= 0; --i) {
     Node& node = *graph.GetNode(node_ids[i]);
 
     if (!SatisfyCondition(node)) {
